refactor: leaderboard and economy (#14)
* refactor: leaderboard and economy

* feat: fix float bug

* fix: ci

* fix: build

* fix: build again

* fix: build again again

+++ src/lib/util/discord-utilities.ts
@@ -1,5 +1,6 @@
 /* eslint-disable typescript-sort-keys/interface */
 import { setTimeout } from 'node:timers';
+import { isTextBasedChannel } from '@sapphire/discord.js-utilities';
 import { isNullishOrEmpty } from '@sapphire/utilities';
 import type { Guild, Message, MessageCreateOptions, MessagePayload, User } from 'discord.js';
 import type { GuildMessage } from '#lib/types';
@@ -87,8 +88,10 @@ export async function sendTemporaryMessage(
 	options: MessageCreateOptions | MessagePayload | string,
 	timeout = 5_000,
 ) {
-	const msg = await message.channel.send(options);
-	setTimeout(async () => msg.delete(), timeout);
+	if (isTextBasedChannel(message.channel)) {
+		const msg = await message.channel.send(options);
+		setTimeout(async () => msg.delete(), timeout);
+	}
 }
 
 export function isUniqueUsername(user: User) {
